+++ modules/admin-ui/src/main/resources/web/admin/common/js/ui/text/_module.ts
@@ -1,5 +1,3 @@
-///<reference path='TextEditorToolbar.ts' />
-///<reference path='TextEditorEditableArea.ts' />
 ///<reference path='TextInput.ts' />
 ///<reference path='AutosizeTextInput.ts' />
 ///<reference path='PasswordInput.ts' />